api/requirements.txt
# Framework Web
flask==3.1.2
flask-cors==6.0.1

# Processamento de Linguagem Natural
nltk==3.9.1
transformers==4.56.0
torch==2.8.0

# Processamento de PDF
PyPDF2==3.0.1

# OpenAI (opcional)
openai==1.102.0

# Utilitários
python-dotenv==1.1.1
numpy==2.3.2
pandas==2.3.2

# Servidor de produção
gunicorn==23.0.0
